+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.2.4] - 2020/05/10
+* Bugfixes (optional scopes handling in the Access Token Response)
+
+## [1.2.3] - 2020/05/06
+* Bugfixes (multiple scopes handling)
+
 ## [1.2.2] - 2020/05/03
 * Added trim for "scope" parameter